Submission
Had to learn a fair amount of adding music bars, set and clear intervals, playing audios, lots of css formatting things to make the music and list/timer go to the sides and stay to the sides. Had to look up calculations and how to make random values and make only 1 decimal point in the celsius using the toFixed function. I worked on a project using side panels so I used that to my advantage when making the content go on the sides. The clock was a little bit annoying to make, mostly with the internal things and using math.floor to remove decimal points. The adding to the list wasn't too hard and changing the temperature was fine. Other than that was a fun but time consuming project.