P07 Submission
I added a dark mode and learned about tweening and transitioning the position and colors of things. Learned about the @media in CSS for styling and grids for making all the parking spots lined up properly. Was a fun project, the CSS and getting the JSON to work properly took a bit of time but I think it came out alright.